move fs to gale/internal
