+++ tools/ghx/actions/custom_action.go
@@ -14,7 +14,7 @@ import (
 
 	"github.com/aweris/gale/internal/config"
 	"github.com/aweris/gale/internal/core"
-	"github.com/aweris/gale/tools/ghx/internal/fs"
+	"github.com/aweris/gale/internal/fs"
 	"github.com/aweris/gale/tools/ghx/internal/log"
 )
