+++ tools/ghx/runner/runner_step.go
@@ -7,8 +7,8 @@ import (
 
 	"github.com/aweris/gale/internal/config"
 	"github.com/aweris/gale/internal/core"
+	"github.com/aweris/gale/internal/fs"
 	"github.com/aweris/gale/tools/ghx/actions"
-	"github.com/aweris/gale/tools/ghx/internal/fs"
 	"github.com/aweris/gale/tools/ghx/internal/log"
 )
